requirements.txt
-e git+git://github.com/ludwigkraatz/django-app-settings.git@a36419af7855612bc8d37c9832e62be22c42bbc3#egg=app-settings-0.1.12
-e git+git://github.com/ludwigkraatz/django-introspective-api.git@4b1760e3a2af1aa9b8dcc9a2c2fdc55a68adbad9#egg=introspective-api-0.1.20
watchdog==0.7.1
PyYaml==3.11
django-pipeline==1.3.24
django-pipeline-compass==0.1.5
isass==0.1.14
pyScss == 1.2.0
-e git+git@github.com:gitpython-developers/GitPython.git@0e9eef45194039af6b8c22edf06cfd7cb106727a#egg=GitPython-0.3.2.RC1